Night and Day Cycle Testing, Bomb, other minor
Hit registration is half implemented. The terrible thing is managing
collision which I need to work on grasping further if I'm going to help
other people through it. The basic functionality of the Arrow projectile
is terms of it lingering is in, I'll most likely add a "Charge up arrow"
type skill along with the second half of registering damage when using a
bow. I imagine this will be significantly harder with Melee, but enemies
are currently killable (just the player is not currently capable of
dealing damage). TOOTLES.