"Power shot" bow attack 3 added, new bug round 2
Added in a hold button to charge attack (the velocity/distance[and
future damage] are effected by button hold length). its pretty hype i
figured some cool stuff out, but found a minor bug.

bug: when swapping between attacks (using a different attack after a
previous attack was made) the first attack is still considered the same
as the previous. I suspect that adding a simple boolean at the end of
each attack to remove this will solve this, but i worry it could also
make a new bug.