Collision update, bug squish part 2
attacks now fire instantaneously even on first button press, even when
switching, and even for the charge attack!

Collision has been updated to properly sort between the currently
implemented enemy, the player, and the player's projectiles. it appears
to be work as intended except...

when shooting south, the generation of the arrow is inaccurate enough to
cause a collision issue with the player. I can try and set player
attacks to ignore player to work around this, but it will have to be in
a later update.