DAMAGE REGISTRATION
Holy fuck i finally got it, after 10,000 years. Hit events are for the
actor instigating the collision [dealing the damage -> i.e. the
projectile, or weapon] and Damage events are used for both [using an
apply dmg node after the hit event fires for the instigating actor, and
separate code in the blueprint of the receiving actor for a event
receive dmg node. Yeah, it was a bit complicated. This combined with the
non-intuitive collision made for a really bad time.]