Item, Orb, and Utility classes
Added an item class that contains a level integer (for the modifier on
the item) and a count (mainly for me since I'll need it for the
inventory), Orb and Utility are both children of it. Orbs have stat1,
stat2, statMod1, statMod2 variables since they can upgrade up to two
stats. The ones that are statx are a string that will contain the name
of the stat, Derek, use whatever names you want for health, attack,
speed, and none (for stat2 in case it only upgrades one stat) when
checking for this in the player class, just let me know so that I can
update it later. The utility class has a Use function that you can
override in the child classes you make by just redeclaring it in the
child. Let me know if that doesn't work (not sure if blueprints supports
it) and we'll think of another way to do it.