Weapon Swap Functionality Complete, Water waves
-Weapon swap functions as follows: Next to a designated area (will be a
weapon rack eventually) swap weapons by hitting the R key, and it will
cycle through the weapons in the following order bow -> sword -> spear
-> staff, back to bow.
only while in the HUB
-Water material has been improved to have a world offset that allows it
to bob up and down, simulating the rising and falling of tide. It's very
subtle.
-A minor bug involving Charged Arrow was resolved in which is no longer
collided with destructible meshes. This has been resolved.